+++ src/custom-jsx-library/index.ts
@@ -1,6 +1,6 @@
 import { jsx, jsxs, Fragment } from "./jsx-runtime";
 import { JSXProps, ElementType } from "./types";
-import { useState } from "custom-jsx-library/core";
-import { render } from "custom-jsx-library/core";
+import { useState } from "custom-jsx-library/core/hooks";
+import { render } from "custom-jsx-library/renderer";
 
 export { jsx, jsxs, Fragment, type JSXProps, type ElementType, useState, render };
